Fix calculation error:
The number of vaccinated people is not the summ of 1st vaccination and 2nd
vaccination, but the number of the 1st vaccinated people contains the completely
vaccinated people.

For the days to go, I always took the complete population as a bases, which
is obviously wrong, because the days to go depend on the people to go, which is
(POPULATION * 0,7) - already vaccinated people.